@@ -3,9 +3,9 @@ import 'package:example/src/common/localization/localization.dart';
 import 'package:example/src/common/util/app_zone.dart';
 import 'package:example/src/common/util/country_picker_state_mixin.dart';
 import 'package:example/src/common/widget/app.dart';
+import 'package:example/src/common/widget/common_bottom_space.dart';
 import 'package:example/src/common/widget/common_header.dart';
 import 'package:example/src/common/widget/common_padding.dart';
-import 'package:example/src/common/widget/common_space_box.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_simple_country_picker/flutter_simple_country_picker.dart';
@@ -141,7 +141,7 @@ class _MobilePreviewState extends State<Preview>
             ),
 
             // --- Withe space --- //
-            const CommonSpaceBox(),
+            const CommonBottomSpacer(),
           ],
         ),
       ),
