@@ -11,7 +11,7 @@ Future<void>? _$initializeApp;
 /// Initializes the app and prepares it for use.
 Future<void> $initializeApp({
   void Function(int progress, String message)? onProgress,
-  FutureOr<void> Function(void dependencies)? onSuccess,
+  Future<void> Function(void dependencies)? onSuccess,
   void Function(Object error, StackTrace stackTrace)? onError,
 }) => _$initializeApp ??= Future<void>(() async {
   late final WidgetsBinding binding;
