@@ -6,12 +6,11 @@
 library;
 
 export 'src/constant/typedef.dart';
-export 'src/controller/countries_controller.dart' show CountriesState;
-export 'src/localization/localization.dart';
+export 'src/controller/country_controller.dart' show CountryState;
+export 'src/localization/country_localizations.dart' show CountryLocalizations;
 export 'src/model/country.dart';
 export 'src/theme/country_picker_theme.dart';
 export 'src/util/country_input_formatter.dart';
-export 'src/widget/countries_scope.dart' show CountriesScope;
 export 'src/widget/country_phone_input.dart';
-export 'src/widget/country_phone_input_expiremental.dart';
+export 'src/widget/country_scope.dart' show CountryScope;
 export 'src/widget/show_country_picker.dart' show showCountryPicker;
