@@ -1,14 +1,30 @@
+/*
+ * Author: Anton Ustinoff <https://github.com/ziqq> | <a.a.ustinoff@gmail.com>
+ * Date: 25 November 2025
+ */
+
+import 'package:meta/meta.dart';
+
+/// Default phone mask
+@internal
+const String kDefaultPhoneMask = '000 000 0000';
+
+/// Default height for [CountryPhoneInput].
+@internal
+const double kDefaultInputHeight = 56;
+
 /// Default flag size
+@internal
 const double kDefaultFlagSize = 25;
 
 /// Default padding
+@internal
 const double kDefaultPadding = 16;
 
 /// Default indent
+@internal
 const double kDefaultIndent = 10;
 
 /// Default radius
-const double kDefaultRadius = 10;
-
-/// Default phone mask
-const String kDefaultPhoneMask = '000 000 0000';
+@internal
+const double kDefaultRadius = 12;
